apps/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.3.7",
    "docling>=2.64.1",
    "fastapi>=0.127.0",
    "langchain>=1.1.3",
    "langchain-chroma>=1.1.0",
    "langchain-huggingface>=1.2.0",
    "matplotlib>=3.10.8",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.2.0",
    "smolagents[openai]>=1.23.0",
    "uvicorn>=0.38.0",
]

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cu126" }]
torchvision = [{ index = "pytorch-cu126" }]
torchaudio = [{ index = "pytorch-cu126" }]
